Added a new operator to irrString to reduce the char* to stringc comversions. Removed const ref in vector2d set method to avoid conversion to float objects. Minor code improvements regarding type conversions.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code@92 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CAnimatedMeshMD2.cpp
@@ -698,8 +698,8 @@ bool CAnimatedMeshMD2::loadFile(io::IReadFile* file)
 	// create indices
 
 	Indices.reallocate(header.numVertices);
-	s32 count = header.numTriangles*3;
-	for (s32 n=0; n<count; n+=3)
+	s16 count = header.numTriangles*3;
+	for (s16 n=0; n<count; n+=3)
 	{
 		Indices.push_back(n);
 		Indices.push_back(n+1);
